+++ static/js/nzbhydra.js
@@ -834,7 +834,7 @@ angular
 function hydralog() {
     controller.$inject = ["$scope", "$http", "$sce"];
     return {
-        template: '<div cg-busy="{promise:logPromise,message:\'Loading log file\'}"><pre ng-bind-html="log" style="text-align: left"></pre></div>',
+        template: '<div cg-busy="{promise:logPromise,message:\'Loading log file\'}"><pre ng-bind-html="log" style="text-align: left; height: 65vh; overflow-y: scroll"></pre></div>',
         controller: controller
     };