+++ nzbhydra/search_module.py
@@ -121,6 +121,8 @@ class SearchModule(object):
             urls = self.get_ebook_urls(search_request)
         elif search_request.type == "audiobook":
             urls = self.get_audiobook_urls(search_request)
+        elif search_request.type == "comic":
+            urls = self.get_comic_urls(search_request)
         else:
             urls = self.get_search_urls(search_request)
         queries_execution_result = self.execute_queries(urls, search_request)
@@ -150,6 +152,11 @@ class SearchModule(object):
         return []
 
     def get_audiobook_urls(self, search_request):
+        # to extend
+        # if module doesnt support it possibly use (configurable) size restrictions when searching
+        return []
+
+    def get_comic_urls(self, search_request):
         # to extend
         # if module doesnt support it possibly use (configurable) size restrictions when searching
         return []