+++ changelog.md
@@ -6,6 +6,8 @@ Added: Show log in scrollable area.
 
 Added: Comic category. Searches in newznab comic category if available, otherwise in general ebook category or just for the query in case of raw search engines.
 
+Added: Link to indexer details page for entries in download history.
+
 Fixed: Selecting search results from duplicates / same name rows wouldn't work.
 
 Fixed: Search would ask for auth even if access wasn't restricted.